Camera Fix, 1/2 Ping System, OldMouseClick Added
Sorry Cody I reworked your camera a bit. The angle is fixed, the camera follows the ghost. As it stands now the player can move the npc through the house by using right click pings. However the ray tracing for pings isn't in yet, so the NPC follows all pings. As it stands the player can easily move the NPC through the house. There is a ping bug where they player is able to do multiple pings at once, not sure if we want to keep that. The npc does move to the latest ping. I created a script called "OldClickMove" because Scripts fix for floating through walls wasn't really a fix. Yeah it can float through walls but I see no method of speeding up that movement. Honestly I don't feel like it matters much at all, if the Ghost follows the path its no skin off my back, just a time saver. The Old Click controls are currently on the player.